refactor: small refactor in ranker drivers
