+++ tests/unit/drivers/test_chunk2doc_rank_drivers.py
@@ -1,7 +1,7 @@
 import pytest
 
 from jina import Document
-from jina.drivers.rank import Chunk2DocRankDriver
+from jina.drivers.rank.aggregate import Chunk2DocRankDriver
 from jina.executors.rankers import Chunk2DocRanker
 from jina.proto import jina_pb2
 from jina.types.sets import DocumentSet
